Create cuda-config
