+++ cublas-sys/build.rs
@@ -1,1 +1,4 @@
-fn main() {}
+fn main() {
+    println!("cargo:rustc-link-lib=dylib=cublas");
+    println!("cargo:rerun-if-changed=build.rs");
+}
