+++ cuda-driver-sys/build.rs
@@ -1,1 +1,4 @@
-fn main() {}
+fn main() {
+    println!("cargo:rustc-link-lib=dylib=cuda");
+    println!("cargo:rerun-if-changed=build.rs");
+}
